Fix exit panics
Trying to call `process.exit` within a thread that is being tracked by
`wait_for_thread` causes a panic (cannot mutably borrow app data container).

This was caused by `LuaSchedulerExt::wait_for_thread` holding a
reference to the `ThreadMap` inside of the returned future. I've solved
this by not having `ThreadMap::listen` be an async function, and
directly returning a future which does not borrow `self`.